vstd::pcm

Struct Resource

pub struct Resource<P> { /* private fields */ }

Interface for PCM / Resource Algebra ghost state.

RA-based ghost state is a well-established theory that is especially useful for verifying concurrent code. An introduction to the concept can be found in Iris: Monoids and Invariants as an Orthogonal Basis for Concurrent Reasoning or Iris from the ground up.

To embed the concept into Verus, we:

Use a trait, PCM, to embed the well-formedness laws of a resource algebra

use a “tracked ghost” object, Resource<P> (this page) to represent ownership of a resource.

Most operations are fairly standard, just “translated” from the usual CSL presentation into Verus.

alloc to allocate a resource.

join to combine two resources via P::op, and split, its inverse.

validate to assert the validity of any held resource.

update or update_nondeterministic to perform a frame-preserving update.

The interface also includes a nontrivial extension for working with shared references to resources. Shared resources do not compose in a “separating” way via P::op, but rather, in a “potentially overlapping” way (join_shared). Shared resources can also be used to “help” perform frame-preserving updates, as long as they themselves do not change (update_with_shared).

Examples

See:

Any of the examples in this directory

The source code for the fractional resource library

See also

The “storage protocol” formalism is an even more significant extension with additional capabilities for interacting with shared resources.

VerusSync provides a higher-level “swiss army knife” for building useful ghost resources.

Implementations

impl<P: PCM> Resource<P>

pub uninterp fn value(self) -> P

pub uninterp fn loc(self) -> Loc

pub proof fn alloc(value: P) -> tracked out : Self

requires
value.valid(),
ensures
out.value() == value,

pub proof fn join(tracked self, tracked other: Self) -> tracked out : Self

requires
self.loc() == other.loc(),
ensures
out.loc() == self.loc(),

out.value() == P::op(self.value(), other.value()),

pub proof fn split(tracked self, left: P, right: P) -> tracked out : (Self, Self)

requires
self.value() == P::op(left, right),
ensures
out.0.loc() == self.loc(),

out.1.loc() == self.loc(),

out.0.value() == left,

out.1.value() == right,

pub proof fn create_unit(loc: Loc) -> tracked out : Self

ensures
out.value() == P::unit(),

out.loc() == loc,

pub proof fn validate(tracked &self)

ensures
self.value().valid(),

pub proof fn update(tracked self, new_value: P) -> tracked out : Self

requires
frame_preserving_update(self.value(), new_value),
ensures
out.loc() == self.loc(),

out.value() == new_value,

pub proof fn update_nondeterministic(tracked self, new_values: Set<P>) -> tracked out : Self

requires
frame_preserving_update_nondeterministic(self.value(), new_values),
ensures
out.loc() == self.loc(),

new_values.contains(out.value()),

pub proof fn join_shared<'a>(tracked &'a self, tracked other: &'a Self) -> tracked out : &'a Self

requires
self.loc() == other.loc(),
ensures
out.loc() == self.loc(),

incl(self.value(), out.value()),

incl(other.value(), out.value()),

This is useful when you have two (or more) shared resources and want to learn that they agree, as you can combine this validate, e.g., x.join_shared(y).validate().

pub proof fn join_shared_to_target<'a>(tracked &'a self, tracked other: &'a Self, target: P, ) -> tracked out : &'a Self

requires
self.loc() == other.loc(),

conjunct_shared(self.value(), other.value(), target),
ensures
out.loc() == self.loc(),

out.value() == target,

pub proof fn weaken<'a>(tracked &'a self, target: P) -> tracked out : &'a Self

requires
incl(target, self.value()),
ensures
out.loc() == self.loc(),

out.value() == target,

pub proof fn validate_2(tracked &mut self, tracked other: &Self)

requires
old(self).loc() == other.loc(),
ensures
*self == *old(self),

P::op(self.value(), other.value()).valid(),

pub proof fn update_with_shared(tracked self, tracked other: &Self, new_value: P) -> tracked out : Self

requires
self.loc() == other.loc(),

frame_preserving_update(
    P::op(self.value(), other.value()),
    P::op(new_value, other.value()),
),
ensures
out.loc() == self.loc(),

out.value() == new_value,

If x · y --> x · z is a frame-perserving update, and we have a shared reference to x, we can update the y resource to z.

pub proof fn update_nondeterministic_with_shared(tracked self, tracked other: &Self, new_values: Set<P>, ) -> tracked out : Self

requires
self.loc() == other.loc(),

frame_preserving_update_nondeterministic(
    P::op(self.value(), other.value()),
    set_op(new_values, other.value()),
),
ensures
out.loc() == self.loc(),

new_values.contains(out.value()),