vstd::raw_ptr

Struct PointsTo

pub struct PointsTo<T> { /* private fields */ }

Permission to access possibly-initialized, typed memory.

The associated pointer (points_to.ptr()) is always a valid pointer for constructing a reference to the underlying data. That means it’s always aligned to its type (is_aligned) and is non-null (is_nonnull).

Notes

The invariants on a PointsTo are a little more restrictive than is necessary for all Rust operations you might want to do. For example:

With a null pointer to a ZST, Rust lets you read and write (though not take a reference).

#[derive(Copy, Clone)]
#[repr(align(64))]
struct X { }

fn zst_test() {
    let x_ptr: *mut X = std::ptr::null_mut();

    let x = unsafe { *x_ptr };  // allowed

    let x = X { };
    unsafe { *x_ptr = x; }      // allowed

    let j = unsafe { &*x_ptr }; // not allowed because ptr is null
}

The std::ptr::read_unaligned and std::ptr::write_unaligned don’t require the pointer to be aligned.

Currently, these use-cases aren’t supported because PointsTo enforces both non-nullness and alignment.

Implementations

impl<T> PointsTo<T>

pub open spec fn ptr(&self) -> *mut T

{ self.view().ptr }

The pointer that this permission is associated with.

pub open spec fn opt_value(&self) -> MemContents<T>

{ self.view().opt_value }

The (possibly uninitialized) memory that this permission gives access to.

pub open spec fn is_init(&self) -> bool

{ self.opt_value().is_init() }

Returns true if the permission’s associated memory is initialized.

pub open spec fn is_uninit(&self) -> bool

{ self.opt_value().is_uninit() }

Returns true if the permission’s associated memory is uninitialized.

pub open spec fn value(&self) -> T

recommends
self.is_init(),

{ self.opt_value().value() }

If the permission’s associated memory is initialized, returns the value that the pointer points to. Otherwise, the result is meaningless.

pub proof fn is_nonnull(tracked &self)

ensures
self@.ptr@.addr != 0,

Guarantee that the PointsTo points to a non-null address.

pub proof fn is_aligned(tracked &self)

ensures
self@.ptr@.addr as nat % align_of::<T>() == 0,

Guarantee that the PointsTo points to an aligned address.

pub proof fn leak_contents(tracked &mut self)

ensures
self.ptr() == old(self).ptr(),

self.is_uninit(),

“Forgets” about the value stored behind the pointer. Updates the PointsTo value to MemContents::Uninit. Note that this is a proof function, i.e., it is operationally a no-op in executable code, even on the Rust Abstract Machine. Only the proof-code representation changes.

pub proof fn is_disjoint<S>(tracked &mut self, tracked other: &PointsTo<S>)

ensures
*old(self) == *self,

self.ptr() as int + size_of::<T>() <= other.ptr() as int
    || other.ptr() as int + size_of::<S>() <= self.ptr() as int,

Guarantees that the memory ranges associated with two permissions will not overlap, since you cannot have two permissions to the same memory.

Note: If both S and T are non-zero-sized, then this implies the pointers have distinct addresses.

impl<V> PointsTo<V>

pub proof fn into_raw(tracked self) -> tracked points_to_raw : PointsToRaw

requires
self.is_uninit(),
ensures
points_to_raw.is_range(self.ptr().addr() as int, size_of::<V>() as int),

points_to_raw.provenance() == self.ptr()@.provenance,

Creates a PointsToRaw from a PointsTo<V> with the same provenance and a range corresponding to the address of the PointsTo<V> and size of V, provided that the memory tracked by the PointsTo<V>is uninitialized.