vstd::seq_lib
Function lemma_flatten_alt_concat

pub proof fn lemma_flatten_alt_concat<A>(x: Seq<Seq<A>>, y: Seq<Seq<A>>)

ensures
(x + y).flatten_alt() =~= x.flatten_alt() + y.flatten_alt(),

Flattening sequences of sequences in reverse order is distributive over concatentation. That is, concatenating the flattening of two sequences of sequences in reverse order is the same as flattening the concatenation of two sequences of sequences in reverse order.