vstd::simple_pptr
Struct PointsTo

pub struct PointsTo<V> { /* private fields */ }

A tracked ghost object that gives the user permission to dereference a pointer for reading or writing, or to free the memory at that pointer.

The meaning of a PointsTo object is given by the data in its View object, PointsToData.

See the PPtr documentation for more details.

Implementations

impl<V> PointsTo<V>

pub open spec fn pptr(&self) -> PPtr<V>

{ PPtr(self.addr(), PhantomData) }

pub closed spec fn addr(self) -> usize

pub closed spec fn mem_contents(&self) -> MemContents<V>

pub open spec fn is_init(&self) -> bool

{ self.mem_contents().is_init() }

pub open spec fn is_uninit(&self) -> bool

{ self.mem_contents().is_uninit() }

pub open spec fn value(&self) -> V

recommends
self.is_init(),

{ self.mem_contents().value() }

pub proof fn is_nonnull(tracked &self)

ensures
self.addr() != 0,

Guarantee that the PointsTo points to a non-null address.

pub proof fn leak_contents(tracked &mut self)

ensures
self.pptr() == old(self).pptr(),

self.is_uninit(),

“Forgets” about the value stored behind the pointer. Updates the PointsTo value to MemContents::Uninit. Note that this is a proof function, i.e., it is operationally a no-op in executable code.

pub proof fn is_disjoint<S>(tracked &mut self, tracked other: &PointsTo<S>)

ensures
*old(self) == *self,

self.addr() + size_of::<V>() <= other.addr()
    || other.addr() + size_of::<S>() <= self.addr(),

Guarantees that two distinct PointsTo<V> objects point to disjoint ranges of memory. If both S and V are non-zero-sized, then this also implies the pointers have distinct addresses.

pub proof fn is_distinct<S>(tracked &mut self, tracked other: &PointsTo<S>)

requires
size_of::<V>() != 0,

size_of::<S>() != 0,
ensures
*old(self) == *self,

self.addr() != other.addr(),

Guarantees that two distinct, non-ZST PointsTo<V> objects point to different addresses. This is a corollary of PointsTo::is_disjoint.