vstd::std_specs::range

Function _verus_external_fn_specification_78_Range_32__58__58__32__60__32_Idx_32__62__32__58__58__32_contains

pub unsafe exec fn _verus_external_fn_specification_78_Range_32__58__58__32__60__32_Idx_32__62__32__58__58__32_contains<Idx, U>(
    r: &Range<Idx>,
    i: &U,
) -> bool
where
    Idx: PartialOrd<U> + PartialOrd<Idx>,
    U: ?Sized + PartialOrd<Idx>,

ensures
<Range<Idx> as ContainsSpec<Idx, U>>::obeys_contains() ==> ret == r.contains_spec(i),

Specification for Range::<Idx>::contains Range::contains method is valid and safe to use only when cmp operations are implemented to satisfy obeys_partial_cmp_spec. Specifically, the comparison must be deterministic, and lt (less than) and le (less than or equal to) must define total orders. If using Range::contains with types that do not satisfy obeys_partial_cmp_spec, no spec is provided.