vstd::storage_protocol

Struct StorageResource

pub struct StorageResource<K, V, P> { /* private fields */ }

Interface for “storage protocol” ghost state. This is an extension-slash-variant on the more well-known concept of “PCM” ghost state, which we also have an interface for here. The unique feature of a storage protocol is the ability to use guard to manipulate shared references of ghost state.

Storage protocols are based on Leaf: Modularity for Temporary Sharing in Separation Logic.

The reference for the laws and operations we’re embedding here can be found at: https://github.com/secure-foundations/leaf/blob/9d72b880feb6af0a7e752b2b2a43806a0812ad56/src/guarding/protocol_relational.v

The reference version requires two monoids, the “protocol monoid” and the “base monoid”. In this interface, we fix the base monoid to be of the form Map<K, V>. (with composition of overlapping maps being undefined), which has all the necessary properties. Note that there’s no create_unit (it’s not sound to do this for an arbitrary location unless you already know a protocol was initialized at that location).

For applications, I generally advise using the tokenized_state_machine! system, rather than using this interface directly.

Implementations

impl<K, V, P: Protocol<K, V>> StorageResource<K, V, P>

pub uninterp fn value(self) -> P

pub uninterp fn loc(self) -> Loc

pub proof fn alloc(p: P, tracked s: Map<K, V>) -> tracked out : Self

requires
P::rel(p, s),
ensures
out.value() == p,

pub proof fn join(tracked a: Self, tracked b: Self) -> tracked out : Self

requires
a.loc() == b.loc(),
ensures
out.loc() == a.loc(),

out.value() == P::op(a.value(), b.value()),

pub proof fn split(tracked self, a_value: P, b_value: P) -> tracked out : (Self, Self)

requires
self.value() == P::op(a_value, b_value),
ensures
out.0.loc() == self.loc(),

out.1.loc() == self.loc(),

out.0.value() == a_value,

out.1.value() == b_value,

pub proof fn validate(tracked &self) -> out : (P, Map<K, V>)

ensures
({
    let (q, t) = out;
    P::rel(P::op(self.value(), q), t)
}),

Since inv isn’t closed under inclusion, validity for an element is defined as the inclusion-closure of invariant, i.e., an element is valid if there exists another element x that, added to it, meets the invariant.

pub proof fn exchange(tracked p: Self, tracked s: Map<K, V>, new_p_value: P, new_s_value: Map<K, V>, ) -> tracked out : (Self, Map<K, V>)

requires
exchanges(p.value(), s, new_p_value, new_s_value),
ensures
({
    let (new_p, new_s) = out;
    new_p.loc() == p.loc() && new_p.value() == new_p_value && new_s == new_s_value
}),

Most general kind of update, potentially depositing and withdrawing

pub proof fn deposit(tracked self, tracked base: Map<K, V>, new_value: P) -> tracked out : Self

requires
deposits(self.value(), base, new_value),
ensures
out.loc() == self.loc(),

out.value() == new_value,

pub proof fn withdraw(tracked self, new_value: P, new_base: Map<K, V>) -> tracked out : (Self, Map<K, V>)

requires
withdraws(self.value(), new_value, new_base),
ensures
out.0.loc() == self.loc(),

out.0.value() == new_value,

out.1 == new_base,

pub proof fn update(tracked self, new_value: P) -> tracked out : Self

requires
updates(self.value(), new_value),
ensures
out.loc() == self.loc(),

out.value() == new_value,

“Normal” update, no depositing or withdrawing

pub proof fn exchange_nondeterministic(tracked p: Self, tracked s: Map<K, V>, new_values: Set<(P, Map<K, V>)>, ) -> tracked out : (Self, Map<K, V>)

requires
exchanges_nondeterministic(p.value(), s, new_values),
ensures
({
    let (new_p, new_s) = out;
    new_p.loc() == p.loc() && new_values.contains((new_p.value(), new_s))
}),

pub proof fn guard(tracked p: &Self, s_value: Map<K, V>) -> tracked s : &Map<K, V>

requires
guards(p.value(), s_value),
ensures
s == s_value,

pub proof fn join_shared<'a>(tracked &'a self, tracked other: &'a Self) -> tracked out : &'a Self

requires
self.loc() == other.loc(),
ensures
out.loc() == self.loc(),

incl(self.value(), out.value()),

incl(other.value(), out.value()),

pub proof fn weaken<'a>(tracked &'a self, target: P) -> tracked out : &'a Self

requires
incl(target, self.value()),
ensures
out.loc() == self.loc(),

out.value() == target,

pub proof fn validate_with_shared(tracked &mut self, tracked x: &Self) -> res : (P, Map<K, V>)

requires
old(self).loc() == x.loc(),
ensures
*self == *old(self),

({
    let (q, t) = res;
    { P::rel(P::op(P::op(self.value(), x.value()), q), t) }
}),

pub proof fn exchange_with_shared(tracked p: Self, tracked x: &Self, tracked s: Map<K, V>, new_p_value: P, new_s_value: Map<K, V>, ) -> tracked out : (Self, Map<K, V>)

requires
p.loc() == x.loc(),

exchanges(P::op(p.value(), x.value()), s, P::op(new_p_value, x.value()), new_s_value),
ensures
out.0.loc() == p.loc(),

out.0.value() == new_p_value,

out.1 == new_s_value,

pub proof fn exchange_nondeterministic_with_shared(tracked p: Self, tracked x: &Self, tracked s: Map<K, V>, new_values: Set<(P, Map<K, V>)>, ) -> tracked out : (Self, Map<K, V>)

requires
p.loc() == x.loc(),

exchanges_nondeterministic(P::op(p.value(), x.value()), s, set_op(new_values, x.value())),
ensures
({
    let (new_p, new_s) = out;
    new_p.loc() == p.loc() && new_values.contains((new_p.value(), new_s))
}),

Most general kind of update, potentially depositing and withdrawing