vstd::tokens
Module frac

Structs

Empty

Token that represents the “empty” state of a fractional resource system. See Frac for more information.

Frac

Token that maintains fractional access to some resource. This allows multiple clients to obtain shared references to some resource via borrow.

FracGhost

An implementation of a resource for fractional ownership of a ghost variable.

GhostVar

See GhostVarAuth<T> for more information.

GhostVarAuth

GhostVarAuth<T> is one half of a pair of tokens—the other being GhostVar<T>. These tokens each track a value, and they can only be allocated or updated together. Thus, the pair of tokens always agree on the value, but they can be owned separately.