vstd::tokens::frac
Struct Frac

pub struct Frac<T, const TOTAL: u64 = 2> { /* private fields */ }

Token that maintains fractional access to some resource. This allows multiple clients to obtain shared references to some resource via borrow.

Implementations

impl<T, const TOTAL: u64> Frac<T, TOTAL>

pub closed spec fn id(self) -> Loc

pub closed spec fn resource(self) -> T

pub closed spec fn frac(self) -> int

pub open spec fn valid(self, id: Loc, frac: int) -> bool

{
    &&& self.id() == id
    &&& self.frac() == frac

}

pub proof fn new(tracked v: T) -> tracked result : Self

requires
TOTAL > 0,
ensures
result.frac() == TOTAL as int,

result.resource() == v,

Create a fractional token that maintains shared access to the input resource.

pub proof fn agree(tracked &self, tracked other: &Self)

requires
self.id() == other.id(),
ensures
self.resource() == other.resource(),

Two tokens agree on values of the underlying resource.

pub proof fn split(tracked &mut self, n: int) -> tracked result : Self

requires
0 < n < old(self).frac(),
ensures
result.id() == self.id() == old(self).id(),

self.resource() == old(self).resource(),

result.resource() == old(self).resource(),

self.frac() + result.frac() == old(self).frac(),

result.frac() == n,

Split one token into two tokens whose quantities sum to the original. The returned token has quantity n; the new value of the input token has quantity old(self).frac() - n.

pub proof fn combine(tracked &mut self, tracked other: Self)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self.resource() == old(self).resource(),

self.resource() == other.resource(),

self.frac() == old(self).frac() + other.frac(),

Combine two tokens, summing their quantities.

pub proof fn bounded(tracked &self)

ensures
0 < self.frac() <= TOTAL,

Allowed values for a token’s quantity.

pub proof fn borrow(tracked &self) -> tracked ret : &T

ensures
ret == self.resource(),

Obtain shared access to the underlying resource.

pub proof fn take_resource(tracked self) -> tracked pair : (T, Empty<T, TOTAL>)

requires
self.frac() == TOTAL,
ensures
pair.0 == self.resource(),

pair.1.id() == self.id(),

Reclaim full ownership of the underlying resource.