vstd::tokens::map

Struct GhostPersistentSubmap

pub struct GhostPersistentSubmap<K, V> { /* private fields */ }

A resource representing duplicable client knowledge of a persistent submap.

The existence of a GhostPersistentSubmap implies that:

Those keys will remain in the map, pointing to the same values, forever;

All other GhostSubmap/GhostPointsTo are disjoint subsets of the domain

Implementations

impl<K, V> GhostPersistentSubmap<K, V>

pub open spec fn is_points_to(self) -> bool

{
    &&& self@.len() == 1
    &&& self.dom().finite()
    &&& !self@.is_empty()

}

Checks whether the GhostPersistentSubmap refers to a single key (and thus can be converted to a GhostPersistentPointsTo).

pub closed spec fn id(self) -> Loc

Resource location

pub closed spec fn view(self) -> Map<K, V>

Logically underlying Map

pub open spec fn dom(self) -> Set<K>

{ self@.dom() }

Domain of the GhostPersistentSubmap

pub open spec fn spec_index(self, key: K) -> V

recommends
self.dom().contains(key),

{ self@[key] }

Indexing operation submap[key]

pub proof fn dummy() -> tracked result : GhostPersistentSubmap<K, V>

Instantiate a dummy GhostPersistentSubmap

pub proof fn empty(id: int) -> tracked result : GhostPersistentSubmap<K, V>

ensures
result.id() == id,

result@ == Map::<K, V>::empty(),

Instantiate an empty GhostPersistentSubmap of a particular id

pub proof fn duplicate(tracked &mut self) -> tracked result : GhostPersistentSubmap<K, V>

ensures
self.id() == result.id(),

old(self).id() == self.id(),

old(self)@ == self@,

result@ == self@,

Duplicate the GhostPersistentSubmap

pub proof fn agree(tracked self: &GhostPersistentSubmap<K, V>, tracked auth: &GhostMapAuth<K, V>)

requires
self.id() == auth.id(),
ensures
self@ <= auth@,

Agreement between a GhostPersistentSubmap and a corresponding GhostMapAuth

Verus might not have full context of the GhostMapAuth and a corresponding GhostPersistentSubmap. However, whenever we know that they refer to the same resource (i.e., have matching ids) we can assert that the GhostPersistentSubmap is a submap of the GhostMapAuth.

proof fn test(tracked &auth: GhostMapAuth<int, int>, tracked &sub: GhostPersistentSubmap<int, int>)
    requires
        auth.id() == sub.id(),
        sub.dom().contains(1int),
        sub[1int] == 1int,
    ensures
        auth[1int] == 1int
{
    sub.agree(auth);
    assert(sub@ <= auth@);
    assert(auth[1int] == 1int);
}

pub proof fn combine(tracked &mut self, tracked other: GhostPersistentSubmap<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@.union_prefer_right(other@),

old(self)@.agrees(other@),

Combining two GhostPersistentSubmaps is possible. We consume the input GhostPersistentSubmap and merge it into the first. We also learn that they agreed

pub proof fn combine_points_to(tracked &mut self, tracked other: GhostPersistentPointsTo<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@.insert(other.key(), other.value()),

old(self)@.contains_key(other.key()) ==> old(self)@[other.key()] == other.value(),

Combining a GhostPersistentPointsTo into GhostPersistentSubmap is possible, in a similar way to the way to combine GhostPersistentSubmaps.

pub proof fn disjoint(tracked &mut self, tracked other: &GhostSubmap<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self@.dom().disjoint(other@.dom()),

When we have a GhostPersistentSubmap and a GhostSubmap we can prove that they have disjoint domains.

pub proof fn intersection_agrees(tracked &mut self, tracked other: &GhostPersistentSubmap<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self@.agrees(other@),

When we have two GhostPersistentSubmaps we can prove that they agree on their intersection.

pub proof fn disjoint_points_to(tracked &mut self, tracked other: &GhostPointsTo<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

!self@.contains_key(other.key()),

When we have a GhostPersistentSubmap and a GhostPointsTo we can prove that they are in disjoint domains.

pub proof fn intersection_agrees_points_to(tracked &mut self, tracked other: &GhostPersistentPointsTo<K, V>, )

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self@.contains_key(other.key()) ==> self@[other.key()] == other.value(),

When we have a GhostPersistentSubmap and a GhostPersistentPointsTo, we can prove that either they are disjoint domains or the key-value pair is in the persistent submap.

pub proof fn split(tracked &mut self, s: Set<K>) -> tracked result : GhostPersistentSubmap<K, V>

requires
s <= old(self)@.dom(),
ensures
self.id() == old(self).id(),

result.id() == self.id(),

old(self)@ == self@.union_prefer_right(result@),

result@.dom() =~= s,

self@.dom() =~= old(self)@.dom() - s,

We can split a GhostPersistentSubmap based on a set of keys in its domain.

pub proof fn split_points_to(tracked &mut self, k: K) -> tracked result : GhostPersistentPointsTo<K, V>

requires
old(self)@.contains_key(k),
ensures
self.id() == old(self).id(),

result.id() == self.id(),

old(self)@ == self@.insert(result.key(), result.value()),

result.key() == k,

self@ == old(self)@.remove(k),

We can separate a single key out of a GhostPersistentSubmap

pub proof fn points_to(tracked self) -> tracked r : GhostPersistentPointsTo<K, V>

requires
self.is_points_to(),
ensures
self@ == map![r.key() => r.value()],

self.id() == r.id(),

Convert a GhostPersistentSubmap into a GhostPersistentPointsTo