vstd::tokens::map

Struct GhostPointsTo

pub struct GhostPointsTo<K, V> { /* private fields */ }

A resource that asserts client ownership over a single key in the map.

The existence of a GhostPointsTo implies that:

Those key will remain in the map (unless the onwer of the GhostPointsTo deletes it using GhostMapAuth::delete_points_to);

Those key will remain pointing to the same value (unless the onwer of the GhostPointsTo updates them using GhostPointsTo::update)

All other GhostSubmap/GhostPointsTo/GhostPersistentSubmap/GhostPersistentPointsTo are disjoint subsets of the domain

Implementations

impl<K, V> GhostPointsTo<K, V>

pub closed spec fn id(self) -> Loc

Location of the underlying resource

pub open spec fn view(self) -> (K, V)

{ (self.key(), self.value()) }

Key-Value pair underlying the points to relationship

pub closed spec fn key(self) -> K

Key of the points to

pub closed spec fn value(self) -> V

Pointed-to value

pub proof fn agree(tracked self: &GhostPointsTo<K, V>, tracked auth: &GhostMapAuth<K, V>)

requires
self.id() == auth.id(),
ensures
auth@.contains_pair(self.key(), self.value()),

Agreement between a GhostPointsTo and a corresponding GhostMapAuth

Verus might not have full context of the GhostMapAuth and a corresponding GhostPointsTo. However, whenever we know that they refer to the same resource (i.e., have matching ids) we can assert that the GhostPointsTo is a key-value pair in the GhostMapAuth.

proof fn test(tracked &auth: GhostMapAuth<int, int>, tracked &pt: GhostPointsTo<int, int>)
    requires
        auth.id() == sub.id(),
        pt@ == (1int, 1int)
    ensures
        auth[1int] == 1int
{
    pt.agree(auth);
    assert(auth[1int] == 1int);
}

pub proof fn combine(tracked self, tracked other: GhostPointsTo<K, V>) -> tracked r : GhostSubmap<K, V>

requires
self.id() == other.id(),
ensures
r.id() == self.id(),

r@ == map![self.key() => self.value(), other.key() => other.value()],

self.key() != other.key(),

We can combine two GhostPointsTos into a GhostSubmap We also learn that they were disjoint.

pub proof fn disjoint(tracked &mut self, tracked other: &GhostPointsTo<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self.key() != other.key(),

Shows that if a two GhostPointsTos are not owning the same key

pub proof fn disjoint_submap(tracked &mut self, tracked other: &GhostSubmap<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

!other.dom().contains(self.key()),

Shows that if a GhostPointsTo and a GhostSubmap are not owning the same key

pub proof fn disjoint_persistent_submap(tracked &mut self, tracked other: &GhostPersistentSubmap<K, V>, )

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

!other.dom().contains(self.key()),

Shows that if a GhostPointsTo and a GhostPersistentSubmap are not owning the same key

pub proof fn disjoint_persistent(tracked &mut self, tracked other: &GhostPersistentPointsTo<K, V>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self.key() != other.key(),

Shows that if a GhostPointsTo and a GhostPersistentPointsTo are not owning the same key

pub proof fn update(tracked &mut self, tracked auth: &mut GhostMapAuth<K, V>, v: V)

requires
old(self).id() == old(auth).id(),
ensures
self.id() == old(self).id(),

auth.id() == old(auth).id(),

self.key() == old(self).key(),

self@ == (self.key(), v),

auth@ == old(auth)@.union_prefer_right(map![self.key() => v]),

Update pointed to value

pub proof fn submap(tracked self) -> tracked r : GhostSubmap<K, V>

ensures
r.id() == self.id(),

r@ == map![self.key() => self.value()],

Convert the GhostPointsTo a GhostSubmap

pub proof fn lemma_view(self)

ensures
self@ == (self.key(), self.value()),

Can be used to learn what the key-value pair of GhostPointsTo is

pub proof fn persist(tracked self) -> tracked r : GhostPersistentPointsTo<K, V>

ensures
self@ == r@,

self.id() == r.id(),

Convert a GhostPointsTo into a GhostPersistentPointsTo