vstd::tokens::set

Struct GhostSubset

pub struct GhostSubset<T> { /* private fields */ }

A resource that has client ownership of a subset

The existence of a GhostSubset implies that:

Those values will remain in the set unless ;

Those values will remain in the set (unless the onwer of the GhostSubset deletes it using GhostSetAuth::delete);

All other GhostSubset/GhostSingleton are disjoint from it

Implementations

impl<T> GhostSubset<T>

pub open spec fn is_singleton(self) -> bool

{
    &&& self@.len() == 1
    &&& self@.finite()
    &&& !self@.is_empty()

}

Checks whether the GhostSubset refers to a single value (and thus can be converted to a GhostSingleton).

pub closed spec fn id(self) -> Loc

Resource location

pub closed spec fn view(self) -> Set<T>

Logically underlying Set

pub proof fn dummy() -> tracked result : GhostSubset<T>

Instantiate a dummy GhostSubset

pub proof fn empty(id: int) -> tracked result : GhostSubset<T>

ensures
result.id() == id,

result@ == Set::<T>::empty(),

Instantiate an empty GhostSubset of a particular id

pub proof fn take(tracked &mut self) -> tracked result : GhostSubset<T>

ensures
old(self).id() == self.id(),

self@.is_empty(),

result == *old(self),

result.id() == self.id(),

Extract the GhostSubset from a mutable reference, leaving behind an empty map.

pub proof fn agree(tracked self: &GhostSubset<T>, tracked auth: &GhostSetAuth<T>)

requires
self.id() == auth.id(),
ensures
self@ <= auth@,

Agreement between a GhostSubset and a corresponding GhostSetAuth

Verus might not have full context of the GhostSetAuth and a corresponding GhostSubset. However, whenever we know that they refer to the same resource (i.e., have matching ids) we can assert that the GhostSubset is a submap of the GhostSetAuth.

proof fn test(tracked &auth: GhostSetAuth<int>, tracked &sub: GhostSubset<int>)
    requires
        auth.id() == sub.id(),
        sub@.contains(1int),
    ensures
        auth@.contains(1int),
{
    sub.agree(auth);
    assert(sub@ <= auth@);
    assert(auth.contains(1int));
}

pub proof fn combine(tracked &mut self, tracked other: GhostSubset<T>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@.union(other@),

old(self)@.disjoint(other@),

Combining two GhostSubsets is possible. We consume the input GhostSubset and merge it into the first. We also learn that they were disjoint.

pub proof fn combine_singleton(tracked &mut self, tracked other: GhostSingleton<T>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@.insert(other@),

!old(self)@.contains(other@),

Combining a GhostSingleton into GhostSubset is possible, in a similar way to the way to combine GhostSubsets.

pub proof fn disjoint(tracked &mut self, tracked other: &GhostSubset<T>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self@.disjoint(other@),

When we have two GhostSubsets we can prove that they have disjoint domains.

pub proof fn disjoint_persistent(tracked &mut self, tracked other: &GhostPersistentSubset<T>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

self@.disjoint(other@),

When we have a GhostSubset and a GhostPersistentSubset we can prove that they are in disjoint domains.

pub proof fn disjoint_singleton(tracked &mut self, tracked other: &GhostSingleton<T>)

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

!self@.contains(other@),

When we have a GhostSubset and a GhostSingleton we can prove that they are in disjoint domains.

pub proof fn disjoint_persistent_singleton(tracked &mut self, tracked other: &GhostPersistentSingleton<T>, )

requires
old(self).id() == other.id(),
ensures
self.id() == old(self).id(),

self@ == old(self)@,

!self@.contains(other@),

When we have a GhostSubset and a GhostPersistentSingleton we can prove that they are in disjoint domains.

pub proof fn split(tracked &mut self, s: Set<T>) -> tracked result : GhostSubset<T>

requires
s <= old(self)@,
ensures
self.id() == old(self).id(),

result.id() == self.id(),

old(self)@ == self@.union(result@),

result@ =~= s,

self@ =~= old(self)@ - s,

We can split a GhostSubset based on a set of values

pub proof fn split_singleton(tracked &mut self, v: T) -> tracked result : GhostSingleton<T>

requires
old(self)@.contains(v),
ensures
self.id() == old(self).id(),

result.id() == self.id(),

old(self)@ == self@.insert(result@),

result@ == v,

self@ =~= old(self)@.remove(v),

We can separate a single value out of a GhostSubset

pub proof fn singleton(tracked self) -> tracked r : GhostSingleton<T>

requires
self.is_singleton(),
ensures
self@ == set![r @],

self.id() == r.id(),

Converting a GhostSubset into a GhostSingleton

pub proof fn persist(tracked self) -> tracked r : GhostPersistentSubset<T>

ensures
self@ == r@,

self.id() == r.id(),

Converting a GhostSubset into a GhostPersistentSubset