vstd

Module wrapping

Contains trusted specifications for wrapping arithmetic behavior on the primitive integer types.

Structured as follows: To get the spec for wrapping_add on u8, for example, call u8_specs::wrapping_add. (The module formulation seemed cleaner than defining a wrapping_add_u8 for every operation and type.)

Modules

i8_specs

i16_specs

i32_specs

i64_specs

i128_specs

isize_specs

u8_specs

u16_specs

u32_specs

u64_specs

u128_specs

usize_specs