vstd
Module atomic_ghost

Provides sequentially-consistent atomic memory locations with associated ghost state. See the atomic_with_ghost! documentation for more information.

Re-exports

pub use atomic_with_ghost_inner;

pub use atomic_with_ghost_store;

pub use atomic_with_ghost_load;

pub use atomic_with_ghost_no_op;

pub use atomic_with_ghost_update_with_1_operand;

pub use atomic_with_ghost_update_with_2_operand;

pub use atomic_with_ghost_update_fetch_add;

pub use atomic_with_ghost_update_fetch_sub;

Macros

atomic_with_ghost

Performs a given atomic operation on a given atomic while providing access to its ghost state.

Structs

AtomicBool

Sequentially-consistent atomic memory location storing a bool and associated ghost state.

AtomicI8

Sequentially-consistent atomic memory location storing a i8 and associated ghost state.

AtomicI16

Sequentially-consistent atomic memory location storing a i16 and associated ghost state.

AtomicI32

Sequentially-consistent atomic memory location storing a i32 and associated ghost state.

AtomicI64

Sequentially-consistent atomic memory location storing a i64 and associated ghost state.

AtomicIsize

Sequentially-consistent atomic memory location storing a isize and associated ghost state.

AtomicPredBool

AtomicPredI8

AtomicPredI16

AtomicPredI32

AtomicPredI64

AtomicPredIsize

AtomicPredPtr

AtomicPredU8

AtomicPredU16

AtomicPredU32

AtomicPredU64

AtomicPredUsize

AtomicPtr

Sequentially-consistent atomic memory location storing a * mut T and associated ghost state.

AtomicU8

Sequentially-consistent atomic memory location storing a u8 and associated ghost state.

AtomicU16

Sequentially-consistent atomic memory location storing a u16 and associated ghost state.

AtomicU32

Sequentially-consistent atomic memory location storing a u32 and associated ghost state.

AtomicU64

Sequentially-consistent atomic memory location storing a u64 and associated ghost state.

AtomicUsize

Sequentially-consistent atomic memory location storing a usize and associated ghost state.

Traits

AtomicInvariantPredicate