vstd
Module tokens

Modules

frac

Structs

InstanceId

Unique identifier for every VerusSync instance. Every “Token” and “Instance” object has an InstanceId. These ID values must agree to perform any token operation.

MapToken

MultisetToken

SetToken

Traits

CountToken

Interface for VerusSync tokens created for a field marked with the count strategy.

ElementToken

Interface for VerusSync tokens created for a field marked with the set, persistent_set or multiset strategies.

KeyValueToken

Interface for VerusSync tokens created for a field marked with the map or persistent_map strategies.

MonotonicCountToken

Interface for VerusSync tokens created for a field marked with the persistent_count strategy.

SimpleToken

Interface for VerusSync tokens created for a field marked with the bool or persistent_bool strategy.

UniqueElementToken

Interface for VerusSync tokens created for a field marked with the set strategy.

UniqueKeyValueToken

Interface for VerusSync tokens created for a field marked with the map strategy.

UniqueSimpleToken

Interface for VerusSync tokens created for a field marked with the bool strategy.

UniqueValueToken

Interface for VerusSync tokens created for a field marked with the variable or option strategies.

ValueToken

Interface for VerusSync tokens created for a field marked with the variable, option or persistent_option strategies.

Functions

bool_value_eq_option_token

bool_value_le_option_token

option_value_eq_option_token

option_value_le_option_token